fix shutil.SameFileError while moving reference log
Refs: https://github.com/Maritsu/spotify-dl/commit/e05d9e444beb6a6e5095aa315b5dcc40c02c8d12#commitcomment-144139425

+++ spotify_dl/youtube.py
@@ -343,8 +343,9 @@ def download_songs(**kwargs):
         log.debug("Downloading to %s", url["save_path"])
     reference_file = DOWNLOAD_LIST
     track_db = write_tracks(reference_file, kwargs["songs"])
-    shutil.copy(reference_file, kwargs["output_dir"] + "/" + reference_file)
-    os.remove(reference_file)
+    if not shutil._samefile(reference_file, kwargs["output_dir"] + "/" + reference_file):
+        shutil.copy(reference_file, kwargs["output_dir"] + "/" + reference_file)
+        os.remove(reference_file)
     reference_file = str(kwargs["output_dir"]) + "/" + reference_file
     kwargs["reference_file"] = reference_file
     kwargs["track_db"] = track_db
